Ranks distribution (#132)
* scripts organization

* updating ranks to work based on int rather than a name and a map of ints and ranks

this way we'd be able to a. fetch them with ease, b. promote with a `user.Rank++` but also to easily find in a map rather than a loop through an array. lastly, users want to knwo rank in int not only a name that doesn't mean much to begin with. One thing to remember here - is that we need to run a query and make sure all the `user.RankName` have been converted to `user.Rank` (int)

* adding rank to uplolad message (#131)

* update upload message and change ranks names and gaps

* adding index ranks to use on production database